Adds the parse.com back-end.
This commit introduces build flavors. The way it's set up is not exactly
elegant; to avoid configuration or dynamic loading I simply have parallel
classes in the two flavors. It seems to work well enough with Gradle and it
means I can build a UIAutomation suite based on stub data. It's unfortunately a
bit awkward in Android Studio; you'll have to unmark one or the other build
flavor from the build path, or it'll complain about duplicate classes.